Flatten serializer into a single file
